Refactor date classes into two separate modules
